Complete removal of core app
