--- asso/core/models/__init__.py
@@ -1,10 +1,0 @@
-from .commons import (
-    Activable,
-    Common,
-    Created,
-    Described,
-    Editable,
-    OrderedModel,
-    Trashable,
-)
-from .users import User
